remove fuse_wait.c
fuse_wait seems to be based on an older macFUSE contract -- fuse_main() does not return until unmount now, while it seems that fuse_wait.c was expecting fusefs_hfs to return. Luckily it did not do much that we need to worry about: after getting a notification that mount was complete, it would basically check to make sure that it worked, but this does not seem totally necessary.